Avoid implicit global var creation and cleanup (#347)
This change will prevent `setopt warn_create_global` from showing errors
and give the reader a better understanding of how the variable is used
(global or local scope).

The preprompt render guard logic of `prompt_pure_cmd_timestamp` was also
removed because this can no longer happen. If a command is running, ZLE
will not be active and `zle reset-prompt` will do nothing.

Closes #345.